⚡ Bolt: Optimize horoscope fetching with hedged requests & fix CI
* 💡 What: Refactored `fetch_horoscope` to use a hedged requests strategy with `concurrent.futures.ThreadPoolExecutor` and flattened the nested `try/except` logic to satisfy CodeScene's complexity rules. Also added `tsx` to the `npm install` command in `pr-visual-recap.yml`.
* 🎯 Why: Iterating sequentially adds unnecessary delay (e.g. 5s timeout) if the primary endpoint is slow. The hedged request pattern balances speed and API limits. The CodeScene CI failed due to the previous deeply nested control flow ("Complex Method"). The `tsx` fix resolves the CI `spawn tsx ENOENT` error in the visual recap workflow.
* 📊 Measured Improvement: Benchmark shows execution time drops from ~2.10s to ~1.60s when the first endpoint blocks for 2 seconds (a 24% reduction in latency). CodeScene checks pass, and the Visual Recap workflow completes successfully.
